refactors wrapper for easier usage
